+++ examples/table-of-contents/batfish.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = () => {
+  return {
+    jsxtremeMarkdownOptions: {
+      wrapper: path.join(__dirname, './src/components/markdown-wrapper.js')
+    }
+  };
+};